+++ CHANGELOG.md
@@ -30,6 +30,7 @@ changes some options. Review your check commands before deploying this version.
 - check-docker-container.rb: output the number of running containers
 - Refactor to connect to the Docker API socket directly instead of using the `docker` or `docker-api` gems
 - Update to rubocop 0.40 and cleanup
+- check-container.rb: issue a critical event if container state 1= running
 
 ## [0.0.4] - 2015-08-10
 ### Changed